This commit changes both JSONToYAML and YAMLToJSON to preserve the map
    key ordering when the document is transformed. This is done by
    reimplementing the library to use YAML v3's yaml.Node API.