@diamondburned 2024-08-12
Benchmark results:

```
goos: linux
goarch: amd64
pkg: github.com/invopop/yaml
cpu: AMD Ryzen 5 5500                               
              │ /tmp/old.bench │           /tmp/new.bench            │
              │     sec/op     │   sec/op     vs base                │
JSONToYAML-12      6.656µ ± 4%   5.811µ ± 1%  -12.69% (p=0.000 n=10)
YAMLToJSON-12      4.688µ ± 3%   4.927µ ± 1%   +5.10% (p=0.001 n=10)
geomean            5.585µ        5.351µ        -4.21%
```

(`old` = before PR; `new` = after PR)