@lzap 2025-04-25
What is the status @samlown I understand you guys patched the security problem and there is that. Do you plan to push new features? I am interested in this and the date parsing problem reported the other day as well. Cheers!