@samlown 2025-04-25
@lzap I honestly didn't have this PR on my radar. We actively use this package in other projects, so our plan is to maintain as much as possible, we're just super-low on time. It looks like @diamondburned did a really good job here, so I'll get it on my Todos.