@samlown 2025-04-25
So this one get a bit more complicated. The go-yaml project on which this is based has been archived 😞  Seems like the best option would be to migrate away to a different solution rather than update to something that is no longer active.