@lzap 2025-04-25
I do not think there is need to migrate off this library, k8s folks already forked it and are fully committed to do security maintenance (no new features): https://github.com/kubernetes-sigs/yaml/tree/master/goyaml.v3 that should be pretty straightforward replacement.