@lzap 2025-04-30
> However, it looks like the underlying project there is a direct fork of the same ghodss/yaml package this one is forked from!

Oh I missed that :-D

Well whatever happens, I am very interested in preserving map ordering, it is a massive pain. @diamondburned would you mind opening this PR at k8s to see how responsive they are? Chances are they might be opinionated on this.